requirements.txt
asgiref==3.2.3
Django==3.0.4
django-annoying==0.10.6
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.3
pendulum==2.1.0
Pillow==7.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
pytzdata==2019.3
six==1.14.0
sqlparse==0.3.1
mysqlclient==1.4.6
